scraper functionality working albeit running it twice
