+++ events/scrapers/generic_crawl4ai.py
@@ -7,90 +7,13 @@ from dotenv import load_dotenv
 from pydantic import BaseModel, Field
 from crawl4ai import AsyncWebCrawler, CrawlerRunConfig, BrowserConfig, CacheMode
 from crawl4ai.extraction_strategy import LLMExtractionStrategy
-from datetime import datetime, timedelta
-import pytz
-import re
+from events.utils.time_parser import format_event_datetime
 
 # Load environment variables from .env file
 load_dotenv()
 
 # Set up logging
 logger = logging.getLogger(__name__)
-
-def parse_datetime(date_str: str, time_str: str) -> tuple[str, str]:
-    """Parse date and time strings into Django's expected format.
-    Returns a tuple of (date, time) strings."""
-    logger.info(f"Parsing date: '{date_str}' and time: '{time_str}'")
-    
-    if not date_str or not time_str:
-        raise ValueError("Date and time strings cannot be empty")
-    
-    try:
-        # Try to parse the date first
-        current_year = datetime.now().year
-        
-        # Clean up the date string
-        date_str = date_str.strip()
-        
-        # If the date string already includes a year, use that
-        if len(date_str.split('/')[-1]) == 4 or len(date_str.split(',')[-1].strip()) == 4:
-            # Date already has a year, use it as is
-            pass
-        else:
-            # No year in the date, need to determine appropriate year
-            try:
-                # Try parsing with current year
-                if '/' in date_str:
-                    test_date = datetime.strptime(f"{date_str}/{current_year}", "%m/%d/%Y")
-                else:
-                    test_date = datetime.strptime(f"{date_str}, {current_year}", "%B %d, %Y")
-                
-                # If the resulting date is more than 1 month in the past, use next year
-                one_month_ago = datetime.now() - timedelta(days=30)
-                if test_date.date() < one_month_ago.date():
-                    current_year += 1
-                    logger.info(f"Date would be in the past, using next year: {current_year}")
-            except ValueError as e:
-                logger.warning(f"Failed to parse date for year check: {str(e)}")
-                raise ValueError(f"Invalid date format: {date_str}")
-        
-        # List of formats to try, in order
-        date_formats = [
-            ("%Y-%m-%d", date_str),                    # YYYY-MM-DD
-            ("%m/%d/%Y", date_str),                    # MM/DD/YYYY
-            ("%B %d, %Y", date_str),                   # Month DD, YYYY
-            ("%m/%d", f"{date_str}/{current_year}"),   # MM/DD -> MM/DD/YYYY
-            ("%B %d", f"{date_str}, {current_year}")   # Month DD -> Month DD, YYYY
-        ]
-        
-        # Try each format until one works
-        date_obj = None
-        last_error = None
-        
-        for fmt, d_str in date_formats:
-            try:
-                date_obj = datetime.strptime(d_str, fmt)
-                break
-            except ValueError as e:
-                last_error = e
-                continue
-        
-        if not date_obj:
-            raise ValueError(f"Could not parse date: {date_str}")
-        
-        # Now parse the time
-        try:
-            # Convert 12-hour time to 24-hour time
-            time_obj = datetime.strptime(time_str, "%I:%M %p")
-            time_str_24h = time_obj.strftime("%H:%M:%S")
-        except ValueError as e:
-            raise ValueError(f"Invalid time format: {time_str}")
-        
-        return date_obj.strftime("%Y-%m-%d"), time_str_24h
-        
-    except Exception as e:
-        logger.error(f"Error parsing date/time: {str(e)}")
-        raise ValueError(f"Invalid date/time format: {str(e)}")
 
 class EventModel(BaseModel):
     event_title: str = Field(..., description="Title of the event.")
@@ -176,42 +99,12 @@ class GenericCrawl4AIScraper:
                 formatted_events = []
                 for event in events_data:
                     try:
-                        # Parse date and time
-                        date, start_time = parse_datetime(
+                        # Use the new time parser utility
+                        start_datetime, end_datetime = format_event_datetime(
                             event.get('event_date', ''),
-                            event.get('event_start_time', '')
-                        )
-                        
-                        # Parse end time if available
-                        _, end_time = parse_datetime(
-                            event.get('event_date', ''),
+                            event.get('event_start_time', ''),
                             event.get('event_end_time', '')
-                        ) if event.get('event_end_time') else ("", "")
-
-                        # Combine date and time for Django format with timezone
-                        tz = pytz.timezone('America/New_York')
-                        start_datetime = None
-                        end_datetime = None
-                        
-                        if date and start_time:
-                            try:
-                                # Create datetime object and make it timezone-aware
-                                dt = datetime.strptime(f"{date} {start_time}", "%Y-%m-%d %H:%M:%S")
-                                start_dt = tz.localize(dt)
-                                start_datetime = start_dt.strftime("%Y-%m-%d %H:%M:%S%z")
-                                
-                                # If no end time provided, set it to 2 hours after start time
-                                if not end_time:
-                                    end_dt = start_dt + timedelta(hours=2)
-                                    end_datetime = end_dt.strftime("%Y-%m-%d %H:%M:%S%z")
-                                else:
-                                    dt = datetime.strptime(f"{date} {end_time}", "%Y-%m-%d %H:%M:%S")
-                                    end_dt = tz.localize(dt)
-                                    end_datetime = end_dt.strftime("%Y-%m-%d %H:%M:%S%z")
-                            except Exception as e:
-                                logger.error(f"Error creating datetime: {str(e)}")
-                                # Don't skip the event, just leave start/end times as None
-                                logger.info("Continuing with null datetime values")
+                        )
 
                         formatted_event = {
                             'title': event.get('event_title', ''),
@@ -228,13 +121,6 @@ class GenericCrawl4AIScraper:
                             'image_url': event.get('event_image_url', ''),
                         }
 
-                        # Only add events that have valid start and end times
-                        #    if start_datetime and end_datetime:
-                        #        logger.info(f"Formatted event: {json.dumps(formatted_event, indent=2)}")
-                        #        formatted_events.append(formatted_event)
-                        #    else:
-                        #        logger.warning(f"Skipping event due to missing datetime: {event.get('event_title')}")
-                        # Add all events, even those without start/end times
                         logger.info(f"Formatted event: {json.dumps(formatted_event, indent=2)}")
                         formatted_events.append(formatted_event)
                     except Exception as e:
